requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.9.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# LLM & LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.50
langgraph-checkpoint>=2.0.0

# API Client
openai>=1.50.0

# UI
streamlit>=1.39.0
plotly>=5.24.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0
